Templated internalizing of WSEGSICD keyword
