+++ opm/parser/eclipse/EclipseState/Schedule/MSW/SICD.hpp
@@ -27,11 +27,10 @@
 #include <string>
 
 #include <opm/parser/eclipse/EclipseState/Schedule/MSW/icd.hpp>
+#include <opm/parser/eclipse/Deck/DeckRecord.hpp>
+#include <opm/parser/eclipse/Deck/DeckKeyword.hpp>
 
 namespace Opm {
-
-    class DeckRecord;
-    class DeckKeyword;
 
     class SICD {
     public:
@@ -89,6 +88,33 @@ namespace Opm {
             serializer(m_max_absolute_rate);
             serializer(m_status);
             serializer(m_scaling_factor);
+        }
+
+        template<class ICD>
+        static std::map<std::string, std::vector<std::pair<int, ICD> > >
+        fromWSEG(const DeckKeyword& wseg) {
+            std::map<std::string, std::vector<std::pair<int, ICD> > > res;
+
+            for (const DeckRecord &record : wseg) {
+                const std::string well_name = record.getItem("WELL").getTrimmedString(0);
+
+                const int start_segment = record.getItem("SEG1").get<int>(0);
+                const int end_segment = record.getItem("SEG2").get<int>(0);
+
+                if (start_segment < 2 || end_segment < 2 || end_segment < start_segment) {
+                    const std::string message = "Segment numbers " + std::to_string(start_segment) + " and "
+                        + std::to_string(end_segment) + " specified in WSEGSICD for well " +
+                        well_name
+                        + " are illegal ";
+                    throw std::invalid_argument(message);
+                }
+
+                const ICD spiral_icd(record);
+                for (int seg = start_segment; seg <= end_segment; seg++) {
+                    res[well_name].push_back(std::make_pair(seg, spiral_icd));
+                }
+            }
+            return res;
         }
 
     private:
